+++ packages/orq-rc/RELEASES.md
@@ -8609,3 +8609,13 @@ Based on:
 - [typescript v4.8.0-rc.54] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.54 - packages/orq-rc
+
+## 2026-04-24 05:15:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.55] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.55 - packages/orq-rc